changed getMedicalRecord(MedicalRecord) methods to return a list

+++ src/main/java/com/openclassrooms/safetynetalerts/service/MedicalRecordService.java
@@ -1,5 +1,8 @@
 package com.openclassrooms.safetynetalerts.service;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,29 +16,36 @@ public class MedicalRecordService {
 	@Autowired
 	private MedicalRecordRepository medicalRecordRepository;
 
-	public MedicalRecord getMedicalRecord(MedicalRecord medicalRecord) {
+	public List<MedicalRecord> getMedicalRecord(MedicalRecord medicalRecord) {
 		return medicalRecordRepository.get(medicalRecord);
 	}
-	
+
 	public MedicalRecord getMedicalRecord(Person person) {
 		if (person == null)
 			return new MedicalRecord();
-		return getMedicalRecord(new MedicalRecord(person.getFirstName(), person.getLastName()));
+		ArrayList<MedicalRecord> medicalRecordsInDB = new ArrayList<MedicalRecord>(
+				getMedicalRecord(new MedicalRecord(person.getFirstName(), person.getLastName())));
+		if (!medicalRecordsInDB.isEmpty())
+			return medicalRecordsInDB.get(0);
+		return new MedicalRecord();
 	}
 
 	public MedicalRecord putMedicalRecord(MedicalRecord medicalRecord) {
 		if (medicalRecord == null)
 			return new MedicalRecord();
 		if (medicalRecord.getFirstName() != null && medicalRecord.getLastName() != null) {
-			MedicalRecord medicalRecordToPut = medicalRecordRepository.get(medicalRecord);
-			if (!medicalRecordToPut.isEmpty()) {
-				medicalRecordRepository.delete(medicalRecordToPut);
-				medicalRecordToPut.update(medicalRecord);
-				return medicalRecordRepository.save(medicalRecordToPut);
-			} else
-				return new MedicalRecord();
-		} else
-			return new MedicalRecord();
+			ArrayList<MedicalRecord> medicalRecordsInDB = new ArrayList<MedicalRecord>(
+					medicalRecordRepository.get(medicalRecord));
+			if (!medicalRecordsInDB.isEmpty()) {
+				MedicalRecord medicalRecordToPut = medicalRecordsInDB.get(0);
+				if (!medicalRecordToPut.isEmpty()) {
+					medicalRecordRepository.delete(medicalRecordToPut);
+					medicalRecordToPut.update(medicalRecord);
+					return medicalRecordRepository.save(medicalRecordToPut);
+				}
+			}
+		}
+		return new MedicalRecord();
 	}
 
 	public MedicalRecord postMedicalRecord(MedicalRecord medicalRecord) {
